rename http-rust-p2 & add http-rust-p3
