--- templates/http-rust-p2/content/src/lib.rs
@@ -1,13 +1,0 @@
-use spin_sdk::http::{IntoResponse, Request, Response};
-use spin_sdk::http_component;
-
-/// A simple Spin HTTP component.
-#[http_component]
-fn handle_{{project-name | snake_case}}(req: Request) -> anyhow::Result<impl IntoResponse> {
-    println!("Handling request to {:?}", req.header("spin-full-url"));
-    Ok(Response::builder()
-        .status(200)
-        .header("content-type", "text/plain")
-        .body("Hello World!")
-        .build())
-}
