+++ templates/http-rust-p3/content/Cargo.toml.tmpl
@@ -3,14 +3,14 @@ name = "{{project-name | kebab_case}}"
 authors = ["{{authors}}"]
 description = "{{project-description}}"
 version = "0.1.0"
-rust-version = "1.78"
-edition = "2021"
+rust-version = "1.93"
+edition = "2024"
 
 [lib]
 crate-type = ["cdylib"]
 
 [dependencies]
 anyhow = "1"
-spin-sdk = "5.2.0"
+spin-sdk = "6.0.0"
 
 [workspace]
