+++ templates/http-rust-p3/content/spin.toml
@@ -13,8 +13,8 @@ route = "{{http-path}}"
 component = "{{project-name | kebab_case}}"
 
 [component.{{project-name | kebab_case}}]
-source = "target/wasm32-wasip1/release/{{project-name | snake_case}}.wasm"
+source = "target/wasm32-wasip2/release/{{project-name | snake_case}}.wasm"
 allowed_outbound_hosts = []
 [component.{{project-name | kebab_case}}.build]
-command = "cargo build --target wasm32-wasip1 --release"
+command = "cargo build --target wasm32-wasip2 --release"
 watch = ["src/**/*.rs", "Cargo.toml"]
