+++ templates/http-rust-p3/content/src/lib.rs
@@ -1,0 +1,12 @@
+use spin_sdk::http::{IntoResponse, Request, Response};
+use spin_sdk::http_service;
+
+/// A simple Spin HTTP component.
+#[http_service]
+async fn handle_{{project-name | snake_case}}(req: Request) -> anyhow::Result<impl IntoResponse> {
+    println!("Handling request to {:?}", req.headers().get("spin-full-url"));
+    Ok(Response::builder()
+        .status(200)
+        .header("content-type", "text/plain")
+        .body("Hello World!".to_string()))
+}
