+++ templates/http-rust-p3/metadata/snippets/component.txt
@@ -3,9 +3,9 @@ route = "{{http-path}}"
 component = "{{project-name | kebab_case}}"
 
 [component.{{project-name | kebab_case}}]
-source = "{{ output-path }}/target/wasm32-wasip1/release/{{project-name | snake_case}}.wasm"
+source = "{{ output-path }}/target/wasm32-wasip2/release/{{project-name | snake_case}}.wasm"
 allowed_outbound_hosts = []
 [component.{{project-name | kebab_case}}.build]
-command = "cargo build --target wasm32-wasip1 --release"
+command = "cargo build --target wasm32-wasip2 --release"
 workdir = "{{ output-path }}"
 watch = ["src/**/*.rs", "Cargo.toml"]
