+++ templates/http-rust/content/src/lib.rs
@@ -1,12 +1,13 @@
 use spin_sdk::http::{IntoResponse, Request, Response};
-use spin_sdk::http_service;
+use spin_sdk::http_component;
 
 /// A simple Spin HTTP component.
-#[http_service]
-async fn handle_{{project-name | snake_case}}(req: Request) -> anyhow::Result<impl IntoResponse> {
-    println!("Handling request to {:?}", req.headers().get("spin-full-url"));
+#[http_component]
+fn handle_{{project-name | snake_case}}(req: Request) -> anyhow::Result<impl IntoResponse> {
+    println!("Handling request to {:?}", req.header("spin-full-url"));
     Ok(Response::builder()
         .status(200)
         .header("content-type", "text/plain")
-        .body("Hello World!".to_string()))
+        .body("Hello World!")
+        .build())
 }
