+++ templates/http-rust/metadata/spin-template.toml
@@ -1,6 +1,6 @@
 manifest_version = "1"
 id = "http-rust"
-description = "HTTP request handler using Rust"
+description = "HTTP request handler compatible with Spin 3.x hosts"
 tags = ["http", "rust"]
 
 [add_component]
